Fikset oversettelse av bildesidetittel. Flyttet funksjon for å sjekke om bilder har endret seg. Fjernet unødvendig sortering.
