+++ src/app/modules/registration/components/edit-images/edit-images.page.html
@@ -4,7 +4,7 @@
       <ion-button (click)="close()">{{ 'CLOSE' | translate }}</ion-button>
     </ion-buttons>
     <ion-title>
-      {{ "REGISTRATION.OVERVIEW.SIMPLE.IMAGES_MODAL_TITLE" | translate: {postfix: modalTitlePostfix | translate | lowercase} }}
+      {{ "REGISTRATION.OVERVIEW.SIMPLE.IMAGES_MODAL_TITLE" | translate: {registrationType: modalTitlePostfix | translate | lowercase} }}
     </ion-title>
   </ion-toolbar>
 </ion-header>
